refactor: Move utilities to pkg helpers
